feat: added guards and secured expense and income modules

+++ src/modules/income/income.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateIncomeDto } from './dto/create-income.dto';
 import { UpdateIncomeDto } from './dto/update-income.dto';
 import { PrismaService } from '../../prisma/prisma.service';
@@ -7,32 +7,52 @@ import { PrismaService } from '../../prisma/prisma.service';
 export class IncomeService {
   constructor(private prisma: PrismaService) {}
 
-  create(createIncomeDto: CreateIncomeDto) {
+  create(createIncomeDto: CreateIncomeDto, userId: string) {
     return this.prisma.income.create({
-      data: { ...createIncomeDto, userId: 'dummy-user-1' },
+      data: { ...createIncomeDto, userId },
     });
   }
 
-  findAll() {
+  findAll(userId: string) {
     return this.prisma.income.findMany({
-      orderBy: { amount: 'desc' },
+      where: { userId },  
+      orderBy: { date: 'desc' },
     });
   }
 
-  findOne(id: string) {
-    return this.prisma.income.findUnique({
-      where: { id },
-    });
+  async findOne(id: string, userId: string) {
+
+    const income = await this.prisma.income.findFirst({
+      where: { id, userId },
+    })
+    if (!income) {
+      throw new NotFoundException('Income not found');
+    }
+    return income;
   }
 
-  update(id: string, updateIncomeDto: UpdateIncomeDto) {
+  async update(id: string, updateIncomeDto: UpdateIncomeDto, userId: string) {
+    const income = await this.prisma.income.findFirst({
+      where: { id, userId },
+    });
+    if (!income) {
+      throw new NotFoundException('Income not found');
+    }
+
     return this.prisma.income.update({
       where: { id },
       data: updateIncomeDto,
     });
   }
 
-  remove(id: string) {
+  async remove(id: string, userId: string) {
+    const income = await this.prisma.income.findFirst({
+      where: { id, userId },
+    });
+    if (!income) {
+      throw new NotFoundException('Income not found');
+    }
+
     return this.prisma.income.delete({
       where: { id },
     });
